requirements.txt
# Discord Bot Framework
discord.py>=2.3.2

# Umgebungsvariablen aus .env laden (optional, aber empfohlen)
python-dotenv>=1.0.0

# Asynchrone HTTP-Anfragen (falls du Webhooks/APIs nutzt)
aiohttp>=3.9.0
jinja2>=3.1.0

# Für Datenbanken (optional)
# asyncpg>=0.29.0          # PostgreSQL
aiosqlite>=0.19.0        # SQLite

# Für JSON-Logging, Configs, etc.
orjson>=3.9.0

# Falls du Pillow für Bilder/Banner brauchst
# Pillow>=10.0.0

# Falls du Voice-Funktionen nutzt
# PyNaCl>=1.5.0
